requirements.txt
google-api-python-client==2.47.0
Flask==2.1.0
gunicorn==20.1.0
functions-framework==3.1.0
slackclient==2.9.4
pandas==1.5.3
numpy==1.24.2
openpyxl==3.1.0